pyproject.toml
[project]
name = "voice-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nomutin", email = "nomura0508@icloud.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aws-lambda-powertools>=3.20.0",
    "einops>=0.8.1",
    "kaldi-native-fbank>=1.22.1",
    "onnxruntime>=1.22.1",
    "soundfile>=0.13.1",
    "soxr>=1.0.0",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = ["COM812", "D", "CPY"]
fixable = ["ALL"]
unfixable = []
exclude = [".git", ".ruff_cache", ".venv",]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
pydocstyle.convention = "numpy"

[tool.mypy]
python_version = "3.13"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[[tool.mypy.overrides]]
module = [
    "kaldi_native_fbank",
    "onnxruntime",
    "soundfile",
    "soxr",
]
ignore_missing_imports = true
